+++ cumulusci/cli/tests/test_cci.py
@@ -14,7 +14,6 @@ from requests.exceptions import ConnectionError
 
 import cumulusci
 from cumulusci.core.config import BaseProjectConfig
-from cumulusci.core.exceptions import ServiceNotConfigured
 from cumulusci.core.exceptions import CumulusCIException
 from cumulusci.cli import cci
 from cumulusci.utils import temporary_dir
@@ -463,450 +462,7 @@ class TestCCI(unittest.TestCase):
         run_click_command(cci.task)
         run_click_command(cci.flow)
         run_click_command(cci.service)
-        run_click_command(cci.service_connect)
         # no assertion; this test is for coverage of empty methods
-
-    @mock.patch("cumulusci.cli.cci.CliTable")
-    def test_service_list__no_active_defaults(self, cli_tbl):
-        runtime = mock.Mock()
-        runtime.project_config.services = {
-            "bad": {"description": "Unconfigured Service"},
-            "test": {"description": "Test Service"},
-            "something_else": {"description": "something else"},
-        }
-        runtime.keychain.list_services.return_value = {
-            "test": ["test_alias", "test2_alias"],
-            "bad": ["bad_alias"],
-        }
-        runtime.keychain._default_services = {"test": "test_alias"}
-        runtime.universal_config.cli__plain_output = None
-
-        run_click_command(
-            cci.service_list, runtime=runtime, plain=False, print_json=False
-        )
-
-        cli_tbl.assert_called_with(
-            [
-                ["Type", "Name", "Default", "Description"],
-                ["bad", "bad_alias", False, "Unconfigured Service"],
-                ["something_else", "", False, "something else"],
-                ["test", "test_alias", True, "Test Service"],
-                ["test", "test2_alias", False, "Test Service"],
-            ],
-            bool_cols=["Default"],
-            dim_rows=[2],
-            title="Services",
-            wrap_cols=["Description"],
-        )
-
-    @mock.patch("cumulusci.cli.cci.CliTable")
-    def test_service_list(self, cli_tbl):
-        runtime = mock.Mock()
-        runtime.project_config.services = {
-            "bad": {"description": "Unconfigured Service"},
-            "test": {"description": "Test Service"},
-            "something_else": {"description": "something else"},
-        }
-        runtime.keychain.list_services.return_value = {
-            "test": ["test_alias", "test2_alias"],
-            "bad": ["bad_alias"],
-        }
-        runtime.keychain._default_services = {"test": "test_alias", "bad": "bad_alias"}
-        runtime.universal_config.cli__plain_output = None
-
-        run_click_command(
-            cci.service_list, runtime=runtime, plain=False, print_json=False
-        )
-
-        cli_tbl.assert_called_with(
-            [
-                ["Type", "Name", "Default", "Description"],
-                ["bad", "bad_alias", True, "Unconfigured Service"],
-                ["something_else", "", False, "something else"],
-                ["test", "test_alias", True, "Test Service"],
-                ["test", "test2_alias", False, "Test Service"],
-            ],
-            bool_cols=["Default"],
-            dim_rows=[2],
-            title="Services",
-            wrap_cols=["Description"],
-        )
-
-    @mock.patch("json.dumps")
-    def test_service_list_json(self, json_):
-        services = {
-            "bad": {"description": "Unconfigured Service"},
-            "test": {"description": "Test Service"},
-        }
-        runtime = mock.Mock()
-        runtime.project_config.services = services
-        runtime.keychain.list_services.return_value = ["test"]
-        runtime.universal_config.cli__plain_output = None
-
-        run_click_command(
-            cci.service_list, runtime=runtime, plain=False, print_json=True
-        )
-
-        json_.assert_called_with(services)
-
-    def test_service_connect__list_commands(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.Mock()
-        runtime.project_config.services = {"test": {}}
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            result = multi_cmd.list_commands(ctx)
-        assert result == ["test"]
-
-    def test_service_connect__list_global_keychain(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.Mock()
-        runtime.project_config = None
-        runtime.universal_config.services = {"test": {}}
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            result = multi_cmd.list_commands(ctx)
-        assert result == ["test"]
-
-    def test_service_connect(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {
-            "test": {"attributes": {"attr": {"required": False}}}
-        }
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test")
-            cmd.callback(ctx.obj, project=True, service_name="test-alias")
-
-            runtime.keychain.set_service.assert_called_once()
-
-            cmd.callback(ctx.obj, project=False, service_name="test-alias")
-
-    @mock.patch("cumulusci.cli.cci.click.confirm")
-    def test_service_connect__alias_already_exists(self, confirm):
-        confirm.side_effect = "y"
-        multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {
-            "test-type": {"attributes": {"attr": {"required": False}}}
-        }
-        runtime.services = {"test-type": {"already-exists": "some config"}}
-        runtime.keychain.list_services.return_value = {"test-type": ["already-exists"]}
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test-type")
-            cmd.callback(
-                runtime,
-                service_type="test-type",
-                service_name="already-exists",
-                project=True,
-            )
-
-        confirm.assert_called_once()
-
-    @mock.patch("click.echo")
-    def test_service_connect__global_default(self, echo):
-        multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {
-            "test": {"attributes": {"attr": {"required": False}}}
-        }
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test")
-            cmd.callback(
-                runtime, service_name="test-alias", default=True, project=False
-            )
-
-        runtime.keychain.set_default_service.assert_called_once_with(
-            "test", "test-alias", project=False
-        )
-        assert (
-            echo.call_args_list[0][0][0] == "Service test:test-alias is now connected"
-        )
-        assert (
-            echo.call_args_list[1][0][0]
-            == "Service test:test-alias is now the default for all CumulusCI projects"
-        )
-
-    @mock.patch("click.echo")
-    def test_service_connect__project_default(self, echo):
-        multi_cmd = cci.ConnectServiceCommand()
-        ctx = mock.Mock()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {
-            "test": {"attributes": {"attr": {"required": False}}}
-        }
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test")
-            cmd.callback(
-                runtime, service_name="test-alias", default=False, project=True
-            )
-
-        runtime.keychain.set_default_service.assert_called_once_with(
-            "test", "test-alias", project=True
-        )
-        assert (
-            echo.call_args_list[0][0][0] == "Service test:test-alias is now connected"
-        )
-        assert (
-            "Service test:test-alias is now the default for project"
-            in echo.call_args_list[1][0][0]
-        )
-
-    def test_service_connect_global_keychain(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.MagicMock()
-        runtime.project_config = None
-        runtime.universal_config.services = {
-            "test": {"attributes": {"attr": {"required": False}}}
-        }
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test")
-            cmd.callback(ctx.obj, project=True, service_name="test-alias")
-
-            runtime.keychain.set_service.assert_called_once()
-
-            cmd.callback(ctx.obj, project=False, service_name="test-alias")
-
-    def test_service_connect_invalid_service(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {}
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            with pytest.raises(click.UsageError):
-                multi_cmd.get_command(ctx, "test")
-
-    def test_service_connect_validator(self):
-        multi_cmd = cci.ConnectServiceCommand()
-        runtime = mock.MagicMock()
-        runtime.project_config.services = {
-            "test": {
-                "attributes": {},
-                "validator": "cumulusci.cli.tests.test_cci.validate_service",
-            }
-        }
-
-        with click.Context(multi_cmd, obj=runtime) as ctx:
-            cmd = multi_cmd.get_command(ctx, "test")
-            with pytest.raises(Exception, match="Validation failed"):
-                cmd.callback(
-                    runtime,
-                    service_type="test",
-                    service_name="test-alias",
-                    project=False,
-                )
-
-    @mock.patch("cumulusci.cli.cci.CliTable")
-    def test_service_info(self, cli_tbl):
-        cli_tbl._table = mock.Mock()
-        service_config = mock.Mock()
-        service_config.config = {"description": "Test Service"}
-        runtime = mock.Mock()
-        runtime.keychain.get_service.return_value = service_config
-        runtime.universal_config.cli__plain_output = None
-
-        run_click_command(
-            cci.service_info,
-            runtime=runtime,
-            service_type="test",
-            service_name="test-alias",
-            plain=False,
-        )
-
-        cli_tbl.assert_called_with(
-            [["Key", "Value"], ["\x1b[1mdescription\x1b[0m", "Test Service"]],
-            title="test/test-alias",
-            wrap_cols=["Value"],
-        )
-
-    @mock.patch("click.echo")
-    def test_service_info_not_configured(self, echo):
-        runtime = mock.Mock()
-        runtime.keychain.get_service.side_effect = ServiceNotConfigured
-
-        run_click_command(
-            cci.service_info,
-            runtime=runtime,
-            service_type="test",
-            service_name="test-alias",
-            plain=False,
-        )
-        assert "not configured for this project" in echo.call_args[0][0]
-
-    @mock.patch("click.echo")
-    def test_service_default__global(self, echo):
-        runtime = mock.Mock()
-        run_click_command(
-            cci.service_default,
-            runtime=runtime,
-            service_type="test",
-            service_name="test-alias",
-            project=False,
-        )
-        runtime.keychain.set_default_service.called_once_with("test", "test-alias")
-        echo.assert_called_once_with(
-            "Service test:test-alias is now the default for all CumulusCI projects"
-        )
-
-    @mock.patch("click.echo")
-    def test_service_default__project(self, echo):
-        runtime = mock.Mock()
-        runtime.keychain.project_local_dir = "test"
-        run_click_command(
-            cci.service_default,
-            runtime=runtime,
-            service_type="test",
-            service_name="test-alias",
-            project=True,
-        )
-        runtime.keychain.set_default_service.called_once_with("test", "test-alias")
-        echo.assert_called_once_with(
-            "Service test:test-alias is now the default for project 'test'"
-        )
-
-    @mock.patch("click.echo")
-    def test_service_default__exception(self, echo):
-        runtime = mock.Mock()
-        runtime.keychain.set_default_service.side_effect = ServiceNotConfigured(
-            "test error"
-        )
-        run_click_command(
-            cci.service_default,
-            runtime=runtime,
-            service_type="no-such-type",
-            service_name="test-alias",
-            project=False,
-        )
-        echo.assert_called_once_with(
-            "An error occurred setting the default service: test error"
-        )
-
-    @mock.patch("click.echo")
-    def test_service_rename(self, echo):
-        runtime = mock.Mock()
-        run_click_command(
-            cci.service_rename,
-            runtime=runtime,
-            service_type="test-type",
-            current_name="old-alias",
-            new_name="new-alias",
-        )
-        runtime.keychain.rename_service.assert_called_once_with(
-            "test-type", "old-alias", "new-alias"
-        )
-        echo.assert_called_once_with(
-            "Service test-type:old-alias has been renamed to new-alias"
-        )
-
-    @mock.patch("click.echo")
-    def test_service_rename__exception(self, echo):
-        runtime = mock.Mock()
-        runtime.keychain.rename_service.side_effect = ServiceNotConfigured("test error")
-        run_click_command(
-            cci.service_rename,
-            runtime=runtime,
-            service_type="test-type",
-            current_name="old-alias",
-            new_name="new-alias",
-        )
-        echo.assert_called_once_with(
-            "An error occurred renaming the service: test error"
-        )
-
-    @mock.patch("cumulusci.cli.cci.click")
-    def test_service_remove(self, click):
-        click.prompt.side_effect = ("future-default-alias",)
-        runtime = mock.Mock()
-        runtime.keychain.services = {
-            "github": {
-                "current-default-alias": "config1",
-                "another-alias": "config2",
-                "future-default-alias": "config3",
-            }
-        }
-        runtime.keychain._default_services = {"github": "current-default-alias"}
-        runtime.keychain.list_services.return_value = {
-            "github": ["current-default-alias", "another-alias", "future-default-alias"]
-        }
-        run_click_command(
-            cci.service_remove,
-            runtime=runtime,
-            service_type="github",
-            service_name="current-default-alias",
-        )
-        runtime.keychain.remove_service.assert_called_once_with(
-            "github", "current-default-alias"
-        )
-        runtime.keychain.set_default_service.assert_called_once_with(
-            "github", "future-default-alias"
-        )
-        assert (
-            click.echo.call_args_list[-1][0][0]
-            == "Service github:current-default-alias has been removed."
-        )
-
-    @mock.patch("cumulusci.cli.cci.click")
-    def test_service_remove__name_does_not_exist(self, click):
-        click.prompt.side_effect = ("this-alias-does-not-exist",)
-        runtime = mock.Mock()
-        runtime.keychain.services = {
-            "github": {
-                "current-default-alias": "config1",
-                "another-alias": "config2",
-                "future-default-alias": "config3",
-            }
-        }
-        runtime.keychain._default_services = {"github": "current-default-alias"}
-        runtime.keychain.list_services.return_value = {
-            "github": ["current-default-alias", "another-alias", "future-default-alias"]
-        }
-        run_click_command(
-            cci.service_remove,
-            runtime=runtime,
-            service_type="github",
-            service_name="current-default-alias",
-        )
-        assert (
-            click.echo.call_args_list[-1][0][0]
-            == "No service of type github with name: this-alias-does-not-exist"
-        )
-        assert runtime.keychain.remove_service.call_count == 0
-        assert runtime.keychain.set_default_service.call_count == 0
-
-    @mock.patch("cumulusci.cli.cci.click")
-    def test_service_remove__exception_thrown(self, click):
-
-        click.prompt.side_effect = ("future-default-alias",)
-        runtime = mock.Mock()
-        runtime.keychain.services = {
-            "github": {
-                "current-default-alias": "config1",
-                "another-alias": "config2",
-                "future-default-alias": "config3",
-            }
-        }
-        runtime.keychain._default_services = {"github": "current-default-alias"}
-        runtime.keychain.list_services.return_value = {
-            "github": ["current-default-alias", "another-alias", "future-default-alias"]
-        }
-        runtime.keychain.remove_service.side_effect = ServiceNotConfigured("test error")
-        run_click_command(
-            cci.service_remove,
-            runtime=runtime,
-            service_type="github",
-            service_name="current-default-alias",
-        )
-        assert (
-            click.echo.call_args_list[-1][0][0]
-            == "An error occurred removing the service: test error"
-        )
 
     @mock.patch(
         "cumulusci.cli.runtime.CliRuntime.get_org",
@@ -944,10 +500,6 @@ class TestCCI(unittest.TestCase):
             cci.main(["cci", "flow", "run", "robot_docs"])
         with mock.patch(rtd, mock_validate_debug(True)):
             cci.main(["cci", "flow", "run", "robot_docs", "--debug"])
-
-
-def validate_service(options):
-    raise Exception("Validation failed")
 
 
 def mock_validate_debug(value):
